修改: 人才培养: 删除考试逻辑修改;

+++ person-training/src/main/java/com/itts/personTraining/controller/ks/admin/KsAdminController.java
@@ -117,6 +117,9 @@ public class KsAdminController {
         if (ksDTO == null) {
             throw new WebException(SYSTEM_NOT_FIND_ERROR);
         }
+        if (ksDTO.getSfxf()) {
+            throw new WebException(ISSUEBATCH_DELETE_ERROR);
+        }
         //更新删除状态
         if (!ksService.delete(ksDTO)) {
             throw new WebException(DELETE_FAIL);
